Moving the storage to a separate directory.
