+++ grow/storage/__init__.py
@@ -1,0 +1,8 @@
+"""References for all pod storage."""
+
+# pylint: disable=wildcard-import, unused-wildcard-import
+from grow.storage.google_storage import *
+from grow.storage.errors import *
+from grow.storage.file_storage import *
+
+AUTO = FileStorage
